+++ biopipen/ns/cellranger_pipeline.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from diot import Diot
+from pipen.utils import mark, is_loading_pipeline
 from pipen_args.procgroup import ProcGroup
 
 if TYPE_CHECKING:
@@ -17,20 +18,32 @@ class CellRangerCountPipeline(ProcGroup):
     """The cellranger count pipeline
 
     Run cellranger count for multiple samples and summarize the metrics.
+
+    Args:
+        input (type=list): The list of lists of fastq files.
+            or the list of comma-separated string of fastq files.
+        ids (type=list): The list of ids for the samples.
     """
-    DEFAULTS = Diot(input=None)
+    DEFAULTS = Diot(input=None, ids=None)
 
     def post_init(self):
         """Check if the input is a list of fastq files"""
-        if (
+        if not is_loading_pipeline() and (
             not isinstance(self.opts.input, (list, tuple))
             or len(self.opts.input) == 0
-            or not isinstance(self.opts.input[0], (list, tuple))
         ):
             raise TypeError(
                 "The input of `CellRangerCountPipeline` should be a list of lists of "
                 "fastq files."
             )
+
+        if isinstance(self.opts.input, (list, tuple)):
+            self.opts.input = [
+                [y.strip() for y in x.split(",")]
+                if isinstance(x, str)
+                else x
+                for x in self.opts.input
+            ]
 
     @ProcGroup.add_proc
     def p_cellranger_count(self) -> Proc:
@@ -38,7 +51,10 @@ class CellRangerCountPipeline(ProcGroup):
         from .cellranger import CellRangerCount as _CellRangerCount
 
         class CellRangerCount(_CellRangerCount):
-            input_data = self.opts.input
+            if self.opts.ids:
+                input_data = list(zip(self.opts.input, self.opts.ids))
+            else:
+                input_data = self.opts.input
 
         return CellRangerCount
 
@@ -58,20 +74,32 @@ class CellRangerVdjPipeline(ProcGroup):
     """The cellranger vdj pipeline
 
     Run cellranger vdj for multiple samples and summarize the metrics.
+
+    Args:
+        input (type=list): The list of lists of fastq files.
+            or the list of comma-separated string of fastq files.
+        ids (type=list): The list of ids for the samples.
     """
-    DEFAULTS = Diot(input=None)
+    DEFAULTS = Diot(input=None, ids=None)
 
     def post_init(self):
         """Check if the input is a list of fastq files"""
-        if (
+        if not is_loading_pipeline() and (
             not isinstance(self.opts.input, (list, tuple))
             or len(self.opts.input) == 0
-            or not isinstance(self.opts.input[0], (list, tuple))
         ):
             raise TypeError(
                 "The input of `CellRangerVdjPipeline` should be a list of lists of "
                 "fastq files."
             )
+
+        if isinstance(self.opts.input, (list, tuple)):
+            self.opts.input = [
+                [y.strip() for y in x.split(",")]
+                if isinstance(x, str)
+                else x
+                for x in self.opts.input
+            ]
 
     @ProcGroup.add_proc
     def p_cellranger_vdj(self) -> Proc:
@@ -79,7 +107,10 @@ class CellRangerVdjPipeline(ProcGroup):
         from .cellranger import CellRangerVdj as _CellRangerVdj
 
         class CellRangerVdj(_CellRangerVdj):
-            input_data = self.opts.input
+            if self.opts.ids:
+                input_data = list(zip(self.opts.input, self.opts.ids))
+            else:
+                input_data = self.opts.input
 
         return CellRangerVdj